requirements.txt
pillow
scikit-learn
numpy
pathlib
matplotlib
pandas